requirement.txt
io.py
re.py
colorama.py